+++ vrtool/orm/io/exporters/section_reliability_exporter.py
@@ -43,7 +43,7 @@ class SectionReliabilityExporter(OrmExporterProtocol):
             _added_assessments.append(
                 AssessmentSectionResults.create(
                     beta=beta_value,
-                    time=float(dom_model.SectionReliability.columns[col_idx]),
+                    time=int(dom_model.SectionReliability.columns[col_idx]),
                     section_data=self._section_data,
                 )
             )
